Fix race condition in ErrProtocolNotSupported
Fixed a race condition that caused the library (and http proxy) to not
recognize ErrProtocolNotSupported, which indicates the vehicle does not
support the newer protocol.

The race condition was caused by code that attempted to handshake with
multiple vehicle subsystems (domains) in parallel. If one subsystem
returned ErrProtocolNotSupported, the pending requests to other
subsystems were canceled. In some cases, the context cancellation error
would be propagated to the down the stack instead of
ErrProtocolNotSupported.

With this fix, the library prioritizes non-context cancellation errors.