Fix panic on inet Send/Close race
The Send method of the Connector interface should be safe to run from a
goroutine, as is done by default when first connecting to a vehicle.
However, if the inet Connector implementation is closed before the
client receives a response, this results in Send writing to a closed
channel and causing a panic.