Fix session info error handling
The vehicle sends unsolicited session info when it detects a possible
desync with the client, and so the client checks for session info on
each message it receives before passing it down the stack. However, the
session info mesasge in question could contain a "Key Not Paired" error,
and the client would continue trying to process session info messages
that contained this error.