Allow non-NIST-P256 TLS server keys
Fixes a regression in a safety check on the TLS key that prevented the
HTTP proxy server from launching.

The intended purpose of the safety check is to verify that the user is
not using the same key as both a TLS server key and a
command-authentication key. The code tries to load the TLS key as a
command-authentication key, and then compares it with the actual
command-authentication key. If the loading step fails because the TLS
key is not a NIST-P256 key, then everything is fine. There's no way it
could be being misued. However, a loading failure was being treated as a
fatal error.